source_code/requirements.txt
absl-py==0.10.0
alabaster==0.7.12
# anaconda-client==1.7.2
# anaconda-navigator==1.9.12
# anaconda-project==0.8.3
argh==0.26.2
asn1crypto==1.3.0
# astroid @ file:///tmp/build/80754af9/astroid_1592495912941/work
astropy==4.0.1.post1
atomicwrites==1.4.0
attrs==19.3.0
# autopep8 @ file:///tmp/build/80754af9/autopep8_1592412889138/work
azure-core==1.8.1
azure-cosmos==4.1.0
azure-storage-blob==12.1.0
Babel==2.8.0
backcall==0.2.0
backports.functools-lru-cache==1.6.1
backports.shutil-get-terminal-size==1.0.0
backports.tempfile==1.0
backports.weakref==1.0.post1
beautifulsoup4==4.9.1
# bitarray @ file:///tmp/build/80754af9/bitarray_1594753876414/work
bkcharts==0.2
bleach==3.1.5
# bokeh @ file:///tmp/build/80754af9/bokeh_1593187602660/work
boto==2.49.0
Bottleneck==1.3.2
brotlipy==0.7.0
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
# cloudpickle @ file:///tmp/build/80754af9/cloudpickle_1594141588948/work
# clyent==1.2.2
colorama==0.4.3
# conda==4.8.3
# conda-build==3.18.11
# conda-package-handling==1.7.0
# conda-verify==3.4.2
contextlib2==0.6.0.post1
cryptography==2.9.2
cycler==0.10.0
# Cython @ file:///tmp/build/80754af9/cython_1594831566883/work
cytoolz==0.10.1
# dask @ file:///tmp/build/80754af9/dask-core_1594156306305/work
decorator==4.4.2
defusedxml==0.6.0
# -e git+git@github.com:facebookresearch/detectron2.git@0af92994bc09a1ea256feb7a76f996cdc3930193#egg=detectron2
# diff-match-patch @ file:///tmp/build/80754af9/diff-match-patch_1594828741838/work
# distributed @ file:///tmp/build/80754af9/distributed_1594750293818/work
docutils==0.16
entrypoints==0.3
et-xmlfile==1.0.1
fastcache==1.1.0
ffmpeg-python==0.2.0
filelock==3.0.12
flake8==3.8.3
Flask==1.1.2
Flask-Caching==1.9.0
fsspec==0.7.4
future==0.18.2
fvcore @ git+https://github.com/facebookresearch/fvcore@1411919d5d351f5c0de4c7becbb5e7bb6c4b5b89
# gevent @ file:///tmp/build/80754af9/gevent_1593009537125/work
glob2==0.7
# gmpy2
# ==2.0.8
google-auth==1.21.3
google-auth-oauthlib==0.4.1
greenlet==0.4.16
grpcio==1.32.0
h5py==2.10.0
HeapDict==1.0.1
# html5lib @ file:///tmp/build/80754af9/html5lib_1593446221756/work
# idna @ file:///tmp/build/80754af9/idna_1593446292537/work
# imageio @ file:///tmp/build/80754af9/imageio_1594161405741/work
imagesize==1.2.0
# importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1593446406207/work
# intervaltree @ file:///tmp/build/80754af9/intervaltree_1594361675072/work
# ipykernel @ file:///tmp/build/80754af9/ipykernel_1594753368510/work/dist/ipykernel-5.3.2-py3-none-any.whl
# ipython @ file:///tmp/build/80754af9/ipython_1593447377063/work
ipython-genutils==0.2.0
ipywidgets==7.5.1
isodate==0.6.0
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
# jedi @ file:///tmp/build/80754af9/jedi_1592841866100/work
jeepney==0.4.3
Jinja2==2.11.2
# joblib @ file:///tmp/build/80754af9/joblib_1594236160679/work
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
# jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1594826976318/work
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyterlab==2.1.5
# jupyterlab-server @ file:///tmp/build/80754af9/jupyterlab_server_1594164409481/work
# keyring @ file:///tmp/build/80754af9/keyring_1593109721646/work
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
libarchive-c==2.9
llvmlite
# ==0.33.0+1.g022ab0f
locket==0.2.0
# lxml @ file:///tmp/build/80754af9/lxml_1594826779415/work
Markdown==3.2.2
MarkupSafe==1.1.1
# matplotlib @ file:///tmp/build/80754af9/matplotlib-base_1592846008246/work
mccabe==0.6.1
mistune==0.8.4
mkl-fft
# ==1.1.0
mkl-random
# ==1.1.1
# mkl-service
# ==2.3.0
mock==4.0.2
more-itertools==8.4.0
mpmath==1.1.0
msgpack==1.0.0
msrest==0.6.19
multipledispatch==0.6.0
# navigator-updater
# ==0.2.1
nbconvert==5.6.1
nbformat==5.0.7
# networkx @ file:///tmp/build/80754af9/networkx_1594377231366/work
# nltk @ file:///tmp/build/80754af9/nltk_1592496090529/work
nose==1.3.7
notebook==6.0.3
numba==0.50.1
numexpr==2.7.1
numpy==1.18.5
# numpydoc @ file:///tmp/build/80754af9/numpydoc_1594166760263/work
oauthlib==3.1.0
olefile==0.46
opencv-python==4.4.0.44
# openpyxl @ file:///tmp/build/80754af9/openpyxl_1594167385094/work
packaging==20.4
# pandas @ file:///tmp/build/80754af9/pandas_1592841659173/work
pandocfilters==1.4.2
parso==0.7.0
partd==1.1.0
path==13.1.0
pathlib2==2.3.5
pathtools==0.1.2
patsy==0.5.1
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
# Pillow @ file:///tmp/build/80754af9/pillow_1594307295532/work
pkginfo==1.5.0.1
pluggy==0.13.1
ply==3.11
portalocker==2.0.0
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.13.0
psutil==5.7.0
ptyprocess==0.6.0
# py @ file:///tmp/build/80754af9/py_1593446248552/work
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycodestyle==2.6.0
pycosat==0.6.3
# pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
# pycurl==7.43.0.5
# pydocstyle @ file:///tmp/build/80754af9/pydocstyle_1592848020240/work
pydot==1.4.1
pyflakes==2.2.0
Pygments==2.6.1
# pylint @ file:///tmp/build/80754af9/pylint_1592496053748/work
# pyodbc
# ===4.0.0-unsupported
# pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1594392929924/work
pyparsing==2.4.7
pyrsistent==0.16.0
PySocks==1.7.1
pytest==5.4.3
python-dateutil==2.8.1
# python-jsonrpc-server @ file:///tmp/build/80754af9/python-jsonrpc-server_1594397536060/work
# python-language-server @ file:///tmp/build/80754af9/python-language-server_1594161909948/work
pytz==2020.1
PyWavelets==1.1.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==19.0.1
QDarkStyle==2.8.1
QtAwesome==0.7.2
# qtconsole @ file:///tmp/build/80754af9/qtconsole_1592848611704/work
QtPy==1.9.0
# regex @ file:///tmp/build/80754af9/regex_1593435547652/work
# requests @ file:///tmp/build/80754af9/requests_1592841827918/work
requests-oauthlib==1.3.0
rope==0.17.0
rsa==4.6
# Rtree==0.9.4
ruamel-yaml==0.15.87
scikit-image==0.16.2
# scikit-learn @ file:///tmp/build/80754af9/scikit-learn_1592502866053/work
# scipy @ file:///tmp/build/80754af9/scipy_1592930511789/work
seaborn==0.10.1
SecretStorage==3.1.2
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
sip
# ==4.19.13
six==1.15.0
snowballstemmer==2.0.0
sortedcollections==1.2.1
sortedcontainers==2.2.2
soupsieve==2.0.1
# Sphinx @ file:///tmp/build/80754af9/sphinx_1594223420021/work
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
# sphinxcontrib-websupport @ file:///tmp/build/80754af9/sphinxcontrib-websupport_1593446360927/work
# spyder @ file:///tmp/build/80754af9/spyder_1594832131358/work
# spyder-kernels @ file:///tmp/build/80754af9/spyder-kernels_1594751551963/work
# SQLAlchemy @ file:///tmp/build/80754af9/sqlalchemy_1593446299926/work
statsmodels==0.11.1
# sympy @ file:///tmp/build/80754af9/sympy_1594236545715/work
tables==3.6.1
tabulate==0.8.7
tblib==1.6.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
# threadpoolctl @ file:///tmp/tmp9twdgx9k/threadpoolctl-2.1.0-py3-none-any.whl
# toml @ file:///tmp/build/80754af9/toml_1592853716807/work
toolz==0.10.0
# torch==1.6.0+cu101
# torchvision==0.7.0+cu101
tornado==6.0.4
# tqdm @ file:///tmp/build/80754af9/tqdm_1593446365756/work
traitlets==4.3.3
# typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1592847887441/work
ujson==1.35
unicodecsv==0.14.1
urllib3==1.25.9
# watchdog @ file:///tmp/build/80754af9/watchdog_1593447344699/work
# wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.11.2
# wurlitzer @ file:///tmp/build/80754af9/wurlitzer_1594753850195/work
xlrd==1.2.0
XlsxWriter==1.2.9
xlwt==1.3.0
xmltodict==0.12.0
yacs==0.1.8
# yapf @ file:///tmp/build/80754af9/yapf_1593528177422/work
zict==2.0.0
zipp==3.1.0
zope.event==4.4
zope.interface==4.7.1
